Rename workers to match DAG naming
